split up vote state
