+++ src/app/features/poll/actions/vote.actions.spec.ts
@@ -27,51 +27,15 @@ describe('Vote Actions', () => {
     });
   });
 
-  describe('Track Vote', () => {
+  describe('Set Vote Option', () => {
     it('Should have correct type', () => {
-      const action = new actions.TrackVote(vote.testPayload);
-      expect(action.type).toBe(actions.TRACK_VOTE);
+      const action = new actions.SetVoteOption('');
+      expect(action.type).toBe(actions.SET_VOTE_OPTION);
     });
 
-    it('Should have correct payload', () => {
-      const action = new actions.TrackVote(vote.testPayload);
-      expect(action.payload).toEqual(vote.testPayload);
-    });
-  });
-
-  describe('Track Vote Succeeded', () => {
-    it('Should have correct type', () => {
-      const action = new actions.TrackVoteSucceeded(vote.testVoteInfo);
-      expect(action.type).toBe(actions.TRACK_VOTE_SUCCEEDED);
-    });
-
-    it('Should have correct payload', () => {
-      const action = new actions.TrackVoteSucceeded(vote.testVoteInfo);
-      expect(action.info).toEqual(vote.testVoteInfo);
-    });
-  });
-
-  describe('Load Vote Info', () => {
-    it('Should have correct type', () => {
-      const action = new actions.LoadVoteInfo('Some UserId');
-      expect(action.type).toBe(actions.LOAD_VOTE_INFO);
-    });
-
-    it('Should have correct payload', () => {
-      const action = new actions.LoadVoteInfo('Some UserId');
-      expect(action.userId).toBe('Some UserId');
-    });
-  });
-
-  describe('Load Vote Info Succeeded', () => {
-    it('Should have correct type', () => {
-      const action = new actions.LoadVoteInfoSucceeded([vote.testVoteInfo]);
-      expect(action.type).toBe(actions.LOAD_VOTE_INFO_SUCCEEDED);
-    });
-
-    it('Should have correct payload', () => {
-      const action = new actions.LoadVoteInfoSucceeded([vote.testVoteInfo]);
-      expect(action.info).toEqual([vote.testVoteInfo]);
+    it('Should have correct option', () => {
+      const action = new actions.SetVoteOption('Option');
+      expect(action.option).toBe('Option');
     });
   });
 });
