+++ src/app/features/poll/reducers/root.reducer.ts
@@ -1,8 +1,10 @@
 import { ActionReducerMap, createFeatureSelector, createSelector } from '@ngrx/store';
 
+import * as fromVoteInfo from './vote-info.reducer';
 import * as fromVote from './vote.reducer';
 
 export interface PollState {
+  voteInfo: fromVoteInfo.State;
   vote: fromVote.State;
 }
 
@@ -11,17 +13,21 @@ export interface State {
 }
 
 export const reducers: ActionReducerMap<PollState, any> = {
+  voteInfo: fromVoteInfo.reducer,
   vote: fromVote.reducer
 };
 
 export const _selectPollState = createFeatureSelector<PollState>('poll');
+export const _selectVoteInfo = createSelector(_selectPollState, state => state.voteInfo);
 export const _selectVote = createSelector(_selectPollState, state => state.vote);
 
-export const { selectEntities: _selectVoteInfo } = fromVote.adapter.getSelectors(_selectVote);
+export const { selectEntities: _selectVoteInfoEntities } = fromVoteInfo.adapter.getSelectors(_selectVoteInfo);
 
+export const _selectVoteOption = createSelector(_selectVote, vote => vote.option);
 
 const pollSelectors = {
-  voteInfo: _selectVoteInfo
+  voteInfo: _selectVoteInfoEntities,
+  selectedOption: _selectVoteOption
 };
 
 export default pollSelectors;
