+++ src/app/features/poll/reducers/vote.reducer.spec.ts
@@ -2,38 +2,14 @@ import * as actions from '../actions/vote.actions';
 
 import { reducer, State } from './vote.reducer';
 
-import { vote } from '../../../test-helpers';
-
 describe('Vote Reducer', () => {
-  it('Should set add an item when TrackVoteSucceeded is dispatched', () => {
+  it('Should set option when SetVoteOption is dispatched', () => {
     const initialState: State = {
-      ids: [],
-      entities: {}
+      option: ''
     };
-    const newState = reducer(initialState, new actions.TrackVoteSucceeded(vote.testVoteInfo));
+    const newState = reducer(initialState, new actions.SetVoteOption('Chipotle'));
     expect(newState).toEqual({
-      ids: [vote.testVoteInfo.pollId],
-      entities: {
-        [vote.testVoteInfo.pollId]: {
-          ...vote.testVoteInfo
-        }
-      }
-    });
-  });
-
-  it('Should add all items when LoadVoteInfoSucceeded is dispatched', () => {
-    const initialState: State = {
-      ids: [],
-      entities: {}
-    };
-    const newState = reducer(initialState, new actions.LoadVoteInfoSucceeded([vote.testVoteInfo]));
-    expect(newState).toEqual({
-      ids: [vote.testVoteInfo.pollId],
-      entities: {
-        [vote.testVoteInfo.pollId]: {
-          ...vote.testVoteInfo
-        }
-      }
+      option: 'Chipotle'
     });
   });
 });
