+++ src/app/features/poll/reducers/vote.reducer.ts
@@ -1,26 +1,20 @@
-import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
-
 import * as actions from '../actions/vote.actions';
 
-export interface VoteInfoEntity {
-  pollId: string;
+export interface State {
   option: string;
-  votedOn: number;
 }
 
-export interface State extends EntityState<VoteInfoEntity> { }
-export const adapter: EntityAdapter<VoteInfoEntity> = createEntityAdapter<VoteInfoEntity>({
-  selectId: (entity) => entity.pollId
-});
-const inititalState = adapter.getInitialState();
+const initialState: State = {
+  option: ''
+};
 
-export function reducer(state = inititalState, action: actions.All): State {
+export function reducer(state = initialState, action: actions.All): State {
   switch (action.type) {
-    case actions.TRACK_VOTE_SUCCEEDED: {
-      return adapter.addOne(action.info, state);
-    }
-    case actions.LOAD_VOTE_INFO_SUCCEEDED: {
-      return adapter.addAll(action.info, state);
+    case actions.SET_VOTE_OPTION: {
+      return {
+        ...state,
+        option: action.option
+      };
     }
     default: {
       return state;
