+++ src/app/reducers/error.reducer.ts
@@ -10,7 +10,7 @@ const initialState: State = {
   message: ''
 };
 
-export function reducer(state: State = initialState, action: actions.All): State {
+export function reducer(state = initialState, action: actions.All): State {
   switch (action.type) {
     case actions.APP_ERROR: {
       return {
